backend/requirements.txt
fastapi
uvicorn[standard]
supabase
pydantic
pydantic-settings
python-multipart
httpx
openai
python-dotenv
requests
retell-sdk
